- Refactor - Refactor the fragments and move them to a own package.
Starting: Im starting all the traslations for. Parallel gonna be working on the firebase implementation for migrating register and login.
